Reduce font size for "last seen" time so it is less prominent.
Previously, this was set to 12.5 px, which is 0.5 px *larger* than
the timestamp for the post itself. At a minimum, the "last seen"
timestamp should be the same size as the "posted" timestamp
for consistency reasons, since they are both time stamps.

But I personally *much* prefer having it be slightly smaller,
since the emphasis should be on when the item was posted.
The user's last-seen time is only supplemental information,
and it's very distracting/annoying to me when it is larger.
In fact, I often find myself looking at the wrong one when
I am in a hurry.