Clean up and fix linkedToFrom logic
When looping, the extracted ID was sometimes undefined due to the link
being relative instead of absolute, therefore the split('/')[4]
technique failed. I have switched to using a regex which should be more
foolproof.

Instead of performing individual GET requests to each linked question,
an API call to the /questions/id/linked endpoint is now made. If a post
is returned there and is in the linked questions list, that post must
reference the current page. This should remove a lot of unnecessary GET
requests that were being performed.

Moved the inline CSS to a sox-linkedToFrom-chevron class.

Added comments to explain logic, as it can be quite hard to follow the
code in this feature! Also added sox.debug lines when checking data from
the API result.

Fixes #389.