Improve sox.helpers.getFromAPI to be more extensible
Changed the getFromAPI method parameters to be an object instead of many
different parameters. The old approach meant a) it was difficult to add
new parameters (e.g. add a sort option), and b) it required the callers
to pass the parameters in with the exact same order. This started to
become tedious.

The actual getFromAPI has also been cleaned up so it is easier to add a
new parameter in future if neccessary.