Use API for editReasonTooltip
I'm not sure why, but previously I was performing a GET request to
*each* post that had a revision and manually extracting the edit comment
from the HTML.

This was unnecessary because it a) required many more page requests that
needed and b) was less efficient as it needed to parse the HTML to
extract the revision comment.

The function now uses a single API query with a heavy filter to only get
the data that is required (post ID, revision comment, revision type).

Part of #391.