Use single API call for standOutDupeCloseMigrated
Previosuly, one API call was being made for *each* post that had a
closed status next to its title.

Thi was unnecessary because it required many more API requests than
needed, and so used up more of the the API quota and bandwidth.

The function will now make a single API call with all the IDs that have
a closed status.

Part of #391.