Implement API caching
Most API requests will now be cached for 3 minutes before they are
requested from the API again. Note: the choice of 3 minutes was fairly
abritrary and might need tweaking in future.

This is to try and reduce the number of useless API calls made, saving
bandwidth and should speed up some SOX features (as they will not need
to wait for a response from the API if the data is already saved
locally).

A major change is that `sox.helpers.getFromAPI` will now essentially
return the 'inner' `items` object, rather than the exact response from
the Stack Exchange API (which had extra 'metadata' such as quota usage,
backoff details, and error messages). This means the function will now
return only an array of items, which will include the cached items, if
applicable.

Caching will be enabled *by default*, as long as multiple IDs are
provided (in an array) and the `featureId` property is sent in the call
to `sox.helpers.getFromAPI`.

I've implemented the caching to be per-feature and per-endpoint. This
means if two different features request to the same endpoint for the
same ID(s), the cache will not be shared. This is because features use
their own specific API filters, so sharing the cache between features
would mean there is potential for the data to not be what was requested.

Part of #391.