copyCommentsLink: don't clone button if comments already expanded
Removes setTimeout added when previously trying to fix this issue.

Adds .comment-text to sox-new-comment event element checks.
This means the event will be fired when individual comments change, for
e.g. when navigating to a link with the comment in the URL which causes
the comments to automatically expand.

Fixes #379.