Replace FontAwesome with sox.sprites.svg
Uses an AJAX GET request to get the SVG file. This is so the file is
cached by the browser. The file contents are appended to the <head>.
Note: adding an <img> tag with the src as a remote SVG file is blocked
on SE, probably due to its CSP.