Fix bug with meta sites
The userscript, before, automatically added the prefix 'meta.' to every site to make an API call with it being the sitename. However, that's not true for non-special sites, where the 'meta' should come after the name and before the 'stackexchange'.

Since API also accepts the name of the site + the meta prefix, here, we replace 'stackexchange' with 'meta' instead of having more complex ways to add it after the name and before 'stackexchange'. If it is a special site, add naturally the 'meta.' prefix.