isQuestionHot: get info from the API
Before, the userscript checked if a question was hot from https://stackexchange.com/hot-questions-for-mobile. However, since mid-March, there's a revision item which says when a question became HNQ. This is also in the API as a comment: "<b>Became Hot Network Question</b> ".

Checking if there's such comment in the post (via revisions-by-ids method) and if its creation_date is shorter or equal to 259200 seconds (3 days) does what was done before.