Wrap shareLinksMarkdown code with setTimeout of 0 & cleanup
It seems like SE has some JS code that populates the input field on
click of the share button, which caused our changes to be immediately
overwritten. A setTimeout of 0 fixes this for now, but ideally there's a
better solution we could use!

Also replace the $.each with $.click's within it with a single $.click
function, and use the current .js-share-link as an anchor to find the
.js-input element.

Fixes #434.