Check if selectionStart is less than selectionEnd and vice versa before the assignment
That fixes a bug. The selectionStart/End are set according to how cursor selects the text. So, if the cursor selects the text from right to left (most usual), then the function will not produce the desired result. In this case, the selectionStart will be the selectionEnd and the selectionEnd will be the selectionStart actually. Checking if selectionStart is less than selectionEnd and assigning accordingly fixes that.