Fix hideWelcomeBackMessage feature
The object passed from `sox.helpers.observe` is a jQuery element, which
can't be used as a normal JS Element. Instead, we can do a live search
if the MutationObserver fires.

This seems to fix the issue, however it is inherently a difficult
feature to test because the banner only appears in specific
circumstances.

Attempts to fix #509.